+++ .gitignore
@@ -1,3 +1,4 @@
+dist/
 node_modules/
 npm-debug.log
