+++ lib/cli/index.js
@@ -1,40 +1,6 @@
 /*eslint-disable no-console */
 
 // This file hooks up on require calls to transpile TypeScript.
-const fs = require('fs');
-const ts = require('typescript');
-const old = require.extensions['.ts'];
-
-require.extensions['.ts'] = function(m, filename) {
-  // If we're in node module, either call the old hook or simply compile the
-  // file without transpilation. We do not touch node_modules/**.
-  // We do touch `angular-cli` files anywhere though.
-  if (!filename.match(/angular-cli/) && filename.match(/node_modules/)) {
-    if (old) {
-      return old(m, filename);
-    }
-    return m._compile(fs.readFileSync(filename), filename);
-  }
-
-  // Node requires all require hooks to be sync.
-  const source = fs.readFileSync(filename).toString();
-
-  try {
-    const result = ts.transpile(source, {
-      target: ts.ScriptTarget.ES5,
-      module: ts.ModuleKind.CommonJs
-    });
-
-    // Send it to node to execute.
-    return m._compile(result, filename);
-  } catch (err) {
-    console.error('Error while running script "' + filename + '":');
-    console.error(err.stack);
-    throw err;
-  }
-};
-
-
 const cli = require('ember-cli/lib/cli');
 const path = require('path');
