+++ packages/ast-tools/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "angular-cli",
+  "version": "1.0.0-beta.11-webpack.2",
+  "description": "CLI tool for Angular",
+  "main": "./index.js",
+  "keywords": [
+    "angular",
+    "cli",
+    "ast",
+    "tool"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/angular/angular-cli.git"
+  },
+  "author": "angular",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/angular/angular-cli/issues"
+  },
+  "homepage": "https://github.com/angular/angular-cli",
+  "dependencies": {
+    "rxjs": "^5.0.0-beta.11",
+    "denodeify": "^1.2.1",
+    "typescript": "^2.0.0"
+  }
+}
