+++ packages/ast-tools/tsconfig.json
@@ -1,24 +1,17 @@
 {
   "compilerOptions": {
     "declaration": true,
-    "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
     "mapRoot": "",
     "module": "commonjs",
     "moduleResolution": "node",
     "noEmitOnError": true,
     "noImplicitAny": true,
-    "outDir": "../dist/",
+    "outDir": "../../dist/ast-tools",
     "rootDir": ".",
     "sourceMap": true,
     "sourceRoot": "/",
     "target": "es5",
-    "lib": ["es6"],
-    "paths": {
-      "ast-tools": [ "../../packages/ast-tools/src" ]
-    }
-  },
-  "includes": [
-    "./custom-typings.d.ts"
-  ]
+    "lib": ["es6"]
+  }
 }
